requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.6
